refactor: split giant components (>1000 lines) into focused modules (#280)
Split 4 large frontend components into parent orchestrators
(≤300 lines each) plus extracted sub-components:

- CharacterImportWizard: 1140 → 189 lines + 8 sub-components + store
- LeftSidebar:         1051 → 187 lines + 8 sub-components
- DialogueLine:         997 → 283 lines + 4 modules
- ProseEditor:          954 → 197 lines + 5 sub-components + hook

Behavior-preserving — all 992 existing tests pass without
modification. Import paths unchanged for all consumers.
