+++ apps/frontend/src/components/CharacterImportWizard/index.ts
@@ -1,0 +1,6 @@
+export { CharacterImportWizard } from "./CharacterImportWizard";
+export type { CharacterImportWizardProps } from "./CharacterImportWizard";
+export type {
+  EditableCharacter,
+  CharacterGroup,
+} from "./CharacterImportWizard";
