+++ apps/frontend/src/components/write-mode/ProseEditor/index.ts
@@ -1,0 +1,3 @@
+export { ProseEditor } from "./ProseEditor";
+export type { ProseEditorRef } from "./ProseEditor";
+export type { LineLayoutMode } from "./ProseEditor";
